[cli] Add tspec/cli
